+++ src/main.rs
@@ -2,6 +2,7 @@ mod auth_cookie;
 mod remodel_api;
 mod remodel_context;
 mod roblox_api;
+mod value;
 
 use std::{
     error::Error,