+++ test-scripts/raw-property-string.lua
@@ -5,3 +5,7 @@ assert(stringValue.ClassName == "StringValue")
 
 local value = remodel.getRawProperty(stringValue, "Value")
 assert(value == "Hello")
+
+remodel.setRawProperty(stringValue, "Value", "String", "Hello, world!")
+local value = remodel.getRawProperty(stringValue, "Value")
+assert(value == "Hello, world!")